refactor: extract shared aggregate and typed-array helpers from leiden algorithm files
docs check acknowledged: internal helper extraction only, no user-facing
feature/language/architecture-table changes.

Impact: 10 functions changed, 27 affected
